Revert "Merge branch 'develop' into feature/UI"
This reverts commit 72a29b82c3c46f503c7e27187239e9a504120404, reversing
changes made to 3eeb9bbc4e9a0167ac5e233d88e4ab416070be6d.

+++ WASD/Assets/Scripts/Character/Attribute.cs
@@ -8,20 +8,20 @@ using UnityEngine;
 public class Attribute
 {
     [SerializeField]
-    private int baseValue;
+    private float baseValue;
     private Sprite icon;
 
     private List<int> modifiers = new List<int>();
 
-    public Attribute(int value, string iconName)
+    public Attribute(float value, string iconName)
     {
         baseValue = value;
         icon = Resources.Load<Sprite>(iconName);
     }
 
-    public int GetValue()
+    public float GetValue()
     {
-        int finalValue = baseValue;
+        float finalValue = baseValue;
         modifiers.ForEach(x => finalValue += x);
         return finalValue;
     }
